Malin

See merge request team-3/idatt1002-2023-3!27

+++ src/main/java/no/ntnu/idatt1002/data/Group.java
@@ -1,7 +1,5 @@
 package no.ntnu.idatt1002.data;
 
-import no.ntnu.idatt1002.dao.Database;
-
 import java.util.ArrayList;
 import java.util.List;
 
@@ -12,7 +10,7 @@ import java.util.List;
  */
 public class Group {
   private final List<User> members = new ArrayList<>(); //The HashMap of Member objects representing the group
-  private final long groupId; //The ID of the group
+  private final int groupId; //The ID of the group
   private String name; //The name of the group
 
   /**
@@ -21,7 +19,7 @@ public class Group {
    * @param groupId the ID of the group
    * @param groupName the name of the group
    */
-  public Group(long groupId, String name) {
+  public Group(int groupId, String name) {
     this.groupId = groupId;
     this.name = name;
   }
@@ -57,7 +55,7 @@ public class Group {
    *
    * @return the ID of the group
    */
-  public long getId() {
+  public int getId() {
     return groupId;
   }
